fix update

+++ update.sh
@@ -2,6 +2,14 @@
 git checkout autoBackLauncher.sh
 git pull
 chmod +x autoBackLauncher.sh
-cd unix_server/
-time make serveur
-mv serv ../serv
+if [ -f "serv" ]
+then
+	echo "No need compile server ..."
+else
+	echo "Need compil ..."
+    cd unix_server/
+    time make serveur
+    mv serv ../serv
+fi
+
+echo "Update finshed !"
